task: refactor Step 1 — move security adapters to user.infrastructure.security (Task 178)
Move SecurityConfig, RpmUserDetails, RpmUserDetailsService,
ProblemDetailAccessDeniedHandler (+ RpmUserDetailsServiceTest) from
iam.auth.infrastructure to iam.user.infrastructure.security. Pure package
move, no behavior change. Updated the two cross-package references:
AuthResource import repointed; SecurityCurrentActorProvider redundant
same-package import removed.

Issue: #178
