+++ ProductSpecification/tasks/178-refactoring-collapse-iam-auth/progress.md
@@ -8,11 +8,11 @@ Type: refactoring
 ## Fix
 
 ### Step 1: Move security adapters auth.infrastructure → user.infrastructure.security
-- [ ] refactor (move SecurityConfig, RpmUserDetails, RpmUserDetailsService, ProblemDetailAccessDeniedHandler + RpmUserDetailsServiceTest; update package + imports)
-- [ ] refactor (cleanup: verify imports/static-analysis)
+- [x] refactor (move SecurityConfig, RpmUserDetails, RpmUserDetailsService, ProblemDetailAccessDeniedHandler + RpmUserDetailsServiceTest; update package + imports)
+- [x] refactor (cleanup: verify imports/static-analysis)
 
 ### Step 2: Move auth web auth.infrastructure.web → user.infrastructure.web
-- [ ] refactor (move AuthResource + LoginRequest/ActivateAccountRequest/CurrentUserResponse/ActivationTokenResponse + AuthResourceTest/ActivateAccountRequestTest; move __files/iam/auth/web → __files/iam/user/web; update package + imports + resource paths)
+- [~] refactor (move AuthResource + LoginRequest/ActivateAccountRequest/CurrentUserResponse/ActivationTokenResponse + AuthResourceTest/ActivateAccountRequestTest; move __files/iam/auth/web → __files/iam/user/web; update package + imports + resource paths)
 - [ ] refactor (cleanup: verify imports/static-analysis)
 
 ### Step 3: Move auth integration tests + fixtures iam.auth → iam.user
