+++ src/main/java/by/iivanov/rpm/iam/user/infrastructure/security/SecurityConfig.java
@@ -1,4 +1,4 @@
-package by.iivanov.rpm.iam.auth.infrastructure;
+package by.iivanov.rpm.iam.user.infrastructure.security;
 
 import by.iivanov.rpm.shared.infrastructure.web.SpaRoutes;
 import io.github.wimdeblauwe.errorhandlingspringbootstarter.UnauthorizedEntryPoint;
